[APPLE-1243] Adding data-verification server.
Added an Xcode "scheme" which
triggers a build of AppCore which
sends Sage data to our data-
verification server.  See

    https://ymedialabs.atlassian.net/wiki/display/APPLE/How+to+see+the+data+we+send+to+Sage

for instructions on how to use it.

Note that, at the moment, some of
the files from BreastCancer aren't
being extracted by the data-
verification server, so if you use this,
you'll have to look at the raw
uploaded files (in a folder on your
desktop). I'll add support for these
files in the next few days (I hope).